Make port settable (and use 2002) fix setup of id/_id fixes to crud; implement collec more tests

+++ js/crud.js
@@ -42,15 +42,20 @@ function csvHeader(fields){
 }
 
 // get model, check for error
-function getModel(entity) {
+function getModel(res, entity) {
     model = dico.getModel(entity)
     if (!model) errors.badRequest(res, 'Invalid entity: "' + entity + '".')
     return model
 }
 
+function getCollection(res, model, name) {
+    if (!model.collecsH[name]) errors.badRequest(res, 'Invalid collection: "' + name + '".')
+    return model.collecsH[name]
+}
+
 // get db store for entity
-function getDb(entity) {
-    return new nedb({ filename: dbpath + entity + '.db', autoload: true })
+function getDb(name) {
+    return new nedb({ filename: dbpath + name + '.db', autoload: true })
 }
 
 function lovFields(model) {
@@ -58,19 +63,21 @@ function lovFields(model) {
 }
 
 // augment the result set by the addition of joined fields on lov tables
+// look up value in this field on lovtable._id
+// replace with lovtable.name (normal tables) or lovtable.text (lov tables)
+// output goes in new field.id + _txt
 function joinResult(res, results, format, fields) {
     if (fields.length == 0) return sendResult(res, results, format)
     let field = fields.shift()
     let txtfld = field.id + '_txt'
+    console.log('join', field.lovtable, field.id, txtfld)
     let db = getDb(field.lovtable)
     db.find({ }, (err, docs) => {
-        //console.log(err, docs)
         if (err) return errors.badRequest(res, 'db error: ' + err)
         for (let rowx = 0; rowx < results.length; rowx++) {
-            let expansion = docs.find(r => r._id == results[rowx][field.id])
-            if (expansion) results[rowx][txtfld] = expansion.text
+            let lookup = docs.find(r => r._id == results[rowx][field.id])
+            if (lookup) results[rowx][txtfld] = lookup.name || lookup.text
         }
-        //console.log(results[0])
         joinResult(res, results, format, fields)
     })
 }
@@ -89,11 +96,9 @@ function sendResult(res, results, format) {
         return res.json(results.length?results[0]:null);
     }
     res.setHeader('_count', nbRecords);
-    if(nbRecords && results[0]._full_count){
-        res.setHeader('_full_count', results[0]._full_count);
-    }else{
-        res.setHeader('_full_count', 0);
-    }
+    res.setHeader('_full_count', nbRecords && format.count ? format.count : 0)
+    // TODO: find a better way to pass the total count
+    if(nbRecords && format.count) results[0]["_full_count"] = format.count
     logger.logCount(results.length || 0);
     return res.json(results);
 }
@@ -108,17 +113,25 @@ function getMany(req, res) {
 
     const entity = req.params.entity,
         format = req.query.format || null,
-        model = getModel(entity)
+        model = getModel(res, entity)
     if (!model) return
 
-    let orderby = {}
-        orderby[model.fields[0].id] = 1
+    //let orderby = { [orderby[model.fields[0].id]]: 1 }
+     let orderby = {}
+         orderby[model.fields[0].id] = 1
+    console.log('get all', model.table || entity, 'order by', orderby)
     
     let csvheader = (format==='csv') ? csvHeader(model.fields) : null,
-        db = getDb(entity)
+        db = getDb(model.table || entity),
+        total_count = 0
+    db.count({}, function(err, n) {
+        if (err) return errors.badRequest(res, 'db error: ' + err)
+        total_count = n
+    })
+    // TODO: make sort case-insensitive
     db.find({}).sort(orderby).exec(function(err, docs) {
         if (err) return errors.badRequest(res, 'db error: ' + err)
-        joinResult(res, docs, { csv: csvheader , single: false }, lovFields(model))
+        joinResult(res, docs, { csv: csvheader , single: false, count: total_count }, lovFields(model))
     })
 }
 
@@ -132,10 +145,11 @@ function getOne(req, res) {
 
     const entity = req.params.entity,
         id = +req.params.id,
-        model = getModel(entity)
+        model = getModel(res, entity)
     if (!model) return
 
-    const db = getDb(entity)
+    console.log('get one', model.table || entity)
+    const db = getDb(model.table || entity)
     db.find({ _id: id }, function(err, docs) {
         if (err) return errors.badRequest(res, 'db error: ' + err)
         joinResult(res, docs, { single: true }, lovFields(model))
@@ -151,7 +165,10 @@ function insertOne(req, res) {
     logger.logReq('INSERT ONE', req)
 
     const entity = req.params.entity,
-        db = getDb(entity)
+        model = getModel(res, entity)
+    if (!model) return
+
+    db = getDb(model.table || entity)
     
     if (db) {
         db.insert(data)
@@ -169,8 +186,10 @@ function updateOne(req, res) {
 
     const entity = req.params.entity,
         id = req.params.id,
-        db = getDb(entity)
-    
+        model = getModel(res, entity)
+    if (!model) return
+
+    db = getDb(model.table || entity)
     if (db) {
         db.update({ _id: id }, data)
     }
@@ -187,7 +206,10 @@ function deleteOne(req, res) {
 
     const entity = req.params.entity,
         id = req.params.id,
-        db = getDb(entity)
+        model = getModel(res, entity)
+    if (!model) return
+    
+    db = getDb(model.table || entity)
     
     if (db) {
         db.remove({ _id: id })
@@ -205,13 +227,12 @@ function lovOne(req, res) {
 
     const entity = req.params.entity,
         fid = req.params.field,
-        model = getModel(entity)
+        model = getModel(res, entity)
     if (!model) return
 
-    let field = model.fields.find(function(f) {
-        return f.id === fid
-    })
+    let field = model.fields.find(function(f) { return f.id === fid })
     if (!field) return errors.badRequest(res, 'Unknown field: ' + fid)
+    console.log('get all', field.lovtable)
 
     let db = getDb(field.lovtable)
     db.find({ }, function(err, docs) {
@@ -229,17 +250,26 @@ function lovOne(req, res) {
 function collecOne(req, res) {
     logger.logReq('GET ONE-COLLEC', req)
 
-
     const entity = req.params.entity,
-        fid = req.params.field,
-        fid = req.params.field,
-        db = getDb(entity)
+        collecId = req.params.collec,
+        pId = parseInt(req.query.id, 10),
+        model = getModel(res, entity),
+        collec = getCollection(res, model, collecId)
+    if (!model || !collec) return
 
-    let f = m.fieldsH[fid]
-        
-    if (db) {
-        // TODO
-    }
+//    let where = { [collec.column]: pId }, 
+//        orderby = { [orderby[collec.fields[0].id]]: 1 }
+    let where = {}, 
+        orderby = {}
+    where[collec.column] = pId
+    orderby[collec.fields[0].id] = 1
+    console.log('get', collec.table || entity, 'where', where, 'order by', orderby)
+    
+    db = getDb(collec.table)
+    db.find(where).sort(orderby).exec(function(err, docs) {
+        if (err) return errors.badRequest(res, 'db error: ' + err)
+        sendResult(res, docs, { })
+    })
 }
